requirements.txt
ddsp==3.5.0
magenta==2.1.3
tensorflow-dataset==4.8.3
django
librosa
mido
numpy
djangorestframework
noisereduce
soundfile
